SCRUM-9 fix crud delete player

+++ app/database/crud.py
@@ -26,8 +26,11 @@ def put_host(db: Session, game: Game, player: Player):
     game.host = player
 
 def delete_player(player: Player, game: Game):
-    game.players.remove(player)
-    player.delete()
+    if not game.started:
+        raise Exception("Cannot leave game that has started")
+    else:
+        game.players.remove(player)
+        player.delete()
 
 def get_game_by_id(db: Session, game_id: int):  
     return db.query(Game).filter(Game.id == game_id).first()
